face-recognition/object_detector/requirements.txt
flask
dlib
pillow
numpy
opencv-python
requests
gunicorn
django-environ
raven
blinker
statsd
six